Create new share trip component
